add more checks for comments

+++ src/repositories/commentsRepository.ts
@@ -6,7 +6,7 @@ import type {
 } from "../types/repositories/commentsRepository";
 import { HTTP_STATUSES } from "../utils";
 import { commentsCollection, ticketCollection, userCollection } from "./db";
-import { Db, ObjectId } from "mongodb";
+import { ObjectId } from "mongodb";
 
 const isUserExist = async (userId: string) => {
 	const user = await userCollection.findOne({ userId });
@@ -42,12 +42,18 @@ export const commentsRepository = {
 	async sendMessage(userId: string, data: SendComment) {
 		const user = await isUserExist(userId);
 
+		if (user.isBannedForChat) {
+			throw new DBError(
+				"Вы не можете отправлять сообщения, так как были заблокированы",
+				HTTP_STATUSES.FORRIBDEN_403,
+			);
+		}
+
 		await findQuestion(data.ticketId, data.questionId);
 
 		const { insertedId } = await commentsCollection.insertOne({
 			...data,
-			firstName: user.firstName,
-			secondName: user.secondName,
+			userId,
 		});
 		const savedMessage = await commentsCollection.findOne(insertedId);
 
@@ -58,11 +64,15 @@ export const commentsRepository = {
 		const { _id, ...rest } = savedMessage;
 		return {
 			commentId: _id,
+			firstName: user.firstName,
+			secondName: user.secondName,
 			...rest,
 		};
 	},
 
-	async getAllComments(questionInfo: GetAllComments) {
+	async getAllComments(userId: string, questionInfo: GetAllComments) {
+		const user = await isUserExist(userId);
+
 		await findQuestion(questionInfo.ticketId, questionInfo.questionId);
 
 		const comments = await commentsCollection.find(questionInfo).toArray();
@@ -71,6 +81,8 @@ export const commentsRepository = {
 			const { _id, ...rest } = comment;
 			return {
 				commentId: _id,
+				firstName: user.firstName,
+				secondName: user.secondName,
 				...rest,
 			};
 		});
@@ -78,18 +90,40 @@ export const commentsRepository = {
 		return formattedComments;
 	},
 
-	async deleteComment(data: DeleteComment) {
-		const deletedComment = await commentsCollection.findOneAndDelete({
+	async deleteComment(userId: string, data: DeleteComment) {
+		const user = await isUserExist(userId);
+
+		if (!ObjectId.isValid(data.commentId)) {
+			throw new DBError(
+				"Неверный формат commentId",
+				HTTP_STATUSES.BAD_REQUEST_400,
+			);
+		}
+
+		const comment = await commentsCollection.findOne({
 			_id: new ObjectId(data.commentId),
 		});
 
-		if (!deletedComment) {
+		if (user.role === "user" && user.userId !== userId) {
+			throw new DBError(
+				"У вас недостаточно прав для удаления комментария",
+				HTTP_STATUSES.FORRIBDEN_403,
+			);
+		}
+
+		if (!comment) {
 			throw new DBError("Комментарий не найден", HTTP_STATUSES.NOT_FOUND_404);
 		}
 
-		const { _id, ...rest } = deletedComment;
+		await commentsCollection.findOneAndDelete({
+			_id: comment._id,
+		});
+
+		const { _id, ...rest } = comment;
 		return {
 			commentId: _id,
+			firstName: user.firstName,
+			secondName: user.secondName,
 			...rest,
 		};
 	},
